Split webdriver actions sequence tests into two files;
MozReview-Commit-ID: IYO7uxcFMzR

Upstreamed from https://bugzilla.mozilla.org/show_bug.cgi?id=1397187 [ci skip]
